mysite/requirements.txt
Django==1.8.2
dj-static==0.0.6
django-crispy-forms-ng==2.0.0
static3==0.6.1
